move webexp check from DashboardViewModel to WidgetHostingService
