+++ tools/Dashboard/DevHome.Dashboard/Services/WidgetHostingService.cs
@@ -2,7 +2,9 @@
 // Licensed under the MIT license.
 
 using System;
+using System.Linq;
 using System.Threading.Tasks;
+using DevHome.Common.Services;
 using DevHome.Dashboard.Helpers;
 using Microsoft.Windows.Widgets.Hosts;
 
@@ -10,11 +12,29 @@ namespace DevHome.Dashboard.Services;
 
 public class WidgetHostingService : IWidgetHostingService
 {
+    private readonly IPackageDeploymentService _packageDeploymentService;
+
     private WidgetHost _widgetHost;
     private WidgetCatalog _widgetCatalog;
 
-    public WidgetHostingService()
+    public WidgetHostingService(IPackageDeploymentService packageDeploymentService)
     {
+        _packageDeploymentService = packageDeploymentService;
+    }
+
+    public bool HasValidWebExperiencePack()
+    {
+        var minSupportedVersion400 = new Version(423, 3800);
+        var minSupportedVersion500 = new Version(523, 3300);
+        var version500 = new Version(500, 0);
+
+        // Ensure the application is installed, and the version is high enough.
+        const string packageFamilyName = "MicrosoftWindows.Client.WebExperience_cw5n1h2txyewy";
+        var packages = _packageDeploymentService.FindPackagesForCurrentUser(
+            packageFamilyName,
+            (minSupportedVersion400, version500),
+            (minSupportedVersion500, null));
+        return packages.Any();
     }
 
     public async Task<WidgetHost> GetWidgetHostAsync()
